Add errors for anon formals in proc defs, unannotated formals in types
