+++ frontend/lib/framework/Makefile.include
@@ -23,8 +23,9 @@ ALL_SRCS += ../frontend/lib/framework/*.h ../frontend/lib/framework/*.cpp
 FRONTEND_FRAMEWORK_SRCS =                           \
            compiler-configuration.cpp               \
            Context.cpp                              \
-           ErrorMessage.cpp                         \
            ErrorBase.cpp                            \
+           error-classes-list.cpp                   \
+           ErrorMessage.cpp                         \
            ErrorWriter.cpp                          \
            ID.cpp                                   \
            Location.cpp                             \
