+++ frontend/lib/parsing/chpl.ypp
@@ -487,7 +487,7 @@
 %type <returnTag>  opt_ret_tag
 %type <astTag> class_tag
 
-%type <uniqueStr> ident_use ident_def reserved_word_ident
+%type <uniqueStr> ident_use ident_def reserved_word_ident formal_ident_def
 %type <uniqueStr> internal_type_ident_def reserved_type_ident_use
 %type <uniqueStr> fn_ident assignop_ident op_ident all_op_name opt_label_ident
 %type <uniqueStr> implements_type_ident implements_type_error_ident
@@ -2572,30 +2572,38 @@ formal:
   named_formal
 ;
 
+formal_ident_def:
+  ident_def
+| TUNDERSCORE
+;
+
 named_formal:
-  opt_formal_intent_tag ident_def opt_colon_formal_type opt_init_expr
+  opt_formal_intent_tag formal_ident_def opt_colon_formal_type opt_init_expr
   {
     $$ = context->buildFormal(@$, $1, $2, $3, $4);
   }
-| pragma_ls opt_formal_intent_tag ident_def opt_colon_formal_type opt_init_expr
+| pragma_ls opt_formal_intent_tag formal_ident_def opt_colon_formal_type
+  opt_init_expr
   {
     $$ = context->buildFormal(@$, $2, $3, $4, $5, true);
   }
-| opt_formal_intent_tag ident_def opt_colon_formal_type formal_var_arg_expr
+| opt_formal_intent_tag formal_ident_def opt_colon_formal_type
+  formal_var_arg_expr
   {
     $$ = context->buildVarArgFormal(@$, $1, $2, $3, $4);
   }
-| pragma_ls opt_formal_intent_tag ident_def opt_colon_formal_type formal_var_arg_expr
+| pragma_ls opt_formal_intent_tag formal_ident_def opt_colon_formal_type
+  formal_var_arg_expr
   {
     $$ = context->buildVarArgFormal(@$, $2, $3, $4, $5, true);
   }
-| opt_formal_intent_tag TLP tuple_var_decl_stmt_inner_ls TRP opt_colon_formal_type
-  opt_init_expr
+| opt_formal_intent_tag TLP tuple_var_decl_stmt_inner_ls TRP
+  opt_colon_formal_type opt_init_expr
   {
     $$ = context->buildTupleFormal(@$, $1, $3, $5, $6);
   }
-| opt_formal_intent_tag TLP tuple_var_decl_stmt_inner_ls TRP opt_colon_formal_type
-  formal_var_arg_expr
+| opt_formal_intent_tag TLP tuple_var_decl_stmt_inner_ls TRP
+  opt_colon_formal_type formal_var_arg_expr
   {
     $$ = ERROR(@$, "variable-length argument may not be grouped in a tuple");
   }
