+++ test/functions/fcf/ExplicitAnonFormalInNamedProc.chpl
@@ -1,0 +1,8 @@
+
+
+
+proc test() {
+  proc foo(_: int, _: int) {}
+}
+test();
+
