+++ test/functions/fcf/Motivators.chpl
@@ -22,10 +22,10 @@ module Motivators {
 
   // Test printing function types, one with named formals and one without.
   proc test3() {
-    type F1 = proc(int, int): int;
+    type F1 = proc(_: int, _: int): int;
     assert(F1:string == "proc(int, int): int");
 
-    type F2 = proc(ref: int, ref: int): int;
+    type F2 = proc(ref _: int, ref _: int): int;
     assert(F2:string == "proc(ref: int, ref: int): int");
 
     type F3 = proc(const x: real, in y: complex): void;
@@ -37,8 +37,8 @@ module Motivators {
   test3();
 
   proc test4() {
-    type F1 = proc(int, int): int;
-    type F2 = proc(x: int, int): int;
+    type F1 = proc(_: int, _: int): int;
+    type F2 = proc(x: int, _: int): int;
     assert(F1 != F2);
   }
   test4();
