+++ test/functions/fcf/UnannotatedFormalInProcType.chpl
@@ -1,0 +1,9 @@
+
+
+
+proc test() {
+  type T = proc(int, int): int;
+  writeln(T);
+}
+test();
+
