+++ test/functions/fcf/UnannotatedFormalInProcType.good
@@ -1,0 +1,3 @@
+UnannotatedFormalInProcType.chpl:4: In function 'test':
+UnannotatedFormalInProcType.chpl:5: error: unannotated formal is ambiguous in this context
+UnannotatedFormalInProcType.chpl:5: error: unannotated formal is ambiguous in this context
